+++ doc/changelog.rst
@@ -4,6 +4,13 @@ Changelog
 0.6.1 (unreleased)
 ------------------
 
+Changes
+~~~~~~~
+
+- The event equations are now taken into account in the
+  determination of the adaptive timestep
+  (`#124 <https://github.com/bluescarni/heyoka/pull/124>`__).
+
 Fix
 ~~~